web scraping/requirements.txt
beautifulsoup4
selenium
requests
pandas
numpy
nltk
scikit-learn
tensorflow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openpyxl